third_party/blink/web_tests/platform/generic/fast/forms/file/file-input-reset-validation-expected.txt

  

This tests that a reset file input with a required value doesn't validate.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS invalid is true
PASS successfullyParsed is true

TEST COMPLETE
